requirements.txt
Flask==2.3.2
python-telegram-bot==20.3
uvicorn==0.22.0
